test: azure build with spack
Tested spack on ubuntu 20.04 and 22.04 (first with build cache)
The build took hours so this likely is not the best approach
for early development. We can engage with the spack team to
ensure the build caches are hits, and for now I will (tomorrow)
work on a build that uses vanilla apt packages (I already
started adding them to the dockerfile and will finish up
and re-build to test tomorrow!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>